refactor: delete unused file and delete comments
