+++ frontend/src/lib/actions/my-page/my-page.ts
@@ -1,7 +1,23 @@
 "use server";
 
 import { getServerAxiosInstance } from "@/lib/api/axios-server";
+import { MyPageEventReturnType } from "@/types/event";
 import { UserType } from "@/types/user";
+
+// Fetch event information
+export const getMyPageEventInfo = async (): Promise<MyPageEventReturnType> => {
+  try {
+    const axiosInstance = await getServerAxiosInstance();
+    const response = await axiosInstance.get(`/me/events`);
+    return response.data;
+  } catch (err) {
+    if (err instanceof Error) {
+      throw new Error(err.message || "Event not found");
+    } else {
+      throw new Error(String(err));
+    }
+  }
+};
 
 // Update user information
 export const updateUserInfo = async (updatedData: {
